# docs/swagger.yaml
basePath: /api
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a learning management system API server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Labpro API
  version: "1.0"
paths:
...
  /courses:
    get:
      description: Get a paginated list of available courses with optional search
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 15, max: 50)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              message:
                type: string
              pagination:
                type: object
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all available courses
      tags:
      - courses
    post:
      consumes:
      - multipart/form-data
      description: Create a new course with title, description, instructor, price,
        topics and thumbnail
      parameters:
      - description: Course title
        in: formData
        name: title
        required: true
        type: string
      - description: Course description
        in: formData
        name: description
        type: string
      - description: Course instructor
        in: formData
        name: instructor
        required: true
        type: string
      - description: Course price
        in: formData
        name: price
        required: true
        type: string
      - collectionFormat: csv
        description: Course topics array
        in: formData
        items:
          type: string
        name: topics
        type: array
      - description: Course thumbnail image
        in: formData
        name: thumbnail
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new course (Admin only)
      tags:
      - admin-courses
  /courses/{courseId}:
    delete:
      description: Delete an existing course and all its associated data
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a course (Admin only)
      tags:
      - admin-courses
    get:
      description: Get detailed information about a specific course
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get course by ID
      tags:
      - courses
    put:
      consumes:
      - multipart/form-data
      description: Update an existing course with new information
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: Course title
        in: formData
        name: title
        type: string
      - description: Course description
        in: formData
        name: description
        type: string
      - description: Course instructor
        in: formData
        name: instructor
        type: string
      - description: Course price
        in: formData
        name: price
        type: string
      - collectionFormat: csv
        description: Course topics array
        in: formData
        items:
          type: string
        name: topics
        type: array
      - description: Course thumbnail image
        in: formData
        name: thumbnail
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Forbidden
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a course (Admin only)
      tags:
      - admin-courses
  /courses/{courseId}/buy:
    post:
      description: Purchase/enroll in a specific course
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Purchase a course
      tags:
      - courses
...
  /courses/my-courses:
    get:
      description: Get a paginated list of courses that the user has purchased/enrolled
        in
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 15, max: 50)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                type: array
              message:
                type: string
              pagination:
                type: object
              status:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              data:
                type: object
              message:
                type: string
              status:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user's enrolled courses
      tags:
      - courses
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"